+++ Plugin/Plugin.vcxproj.filters
@@ -66,7 +66,7 @@
     <ClCompile Include="eventDialog.cpp">
       <Filter>源文件</Filter>
     </ClCompile>
-    <ClCompile Include="DateFormat.cpp">
+    <ClCompile Include="Issue66.cpp">
       <Filter>源文件</Filter>
     </ClCompile>
   </ItemGroup>
